requirements.txt
ctransformers
sentence-transformers 
pinecone-client
langchain-community
langchain
flask
pypdf
langchain-openai
langchain-pinecone
langchain-chroma
python-dotenv

-e .